support ignore through html comment above instance script
