+++ src/compiler/compile/Component.ts
@@ -36,7 +36,7 @@ import { DecodedSourceMap, RawSourceMap } from '@ampproject/remapping/dist/types
 import { clone } from '../utils/clone';
 import compiler_warnings from './compiler_warnings';
 import compiler_errors from './compiler_errors';
-import { extract_svelte_ignore_from_comments } from './utils/extract_svelte_ignore';
+import { extract_ignores_above_position, extract_svelte_ignore_from_comments } from './utils/extract_svelte_ignore';
 
 interface ComponentOptions {
 	namespace?: string;
@@ -172,12 +172,17 @@ export default class Component {
 		}
 
 		this.walk_module_js();
+
+		this.push_ignores(this.ast.instance ? extract_ignores_above_position(this.ast.instance.start, this.ast.html.children) : []);
 		this.walk_instance_js_pre_template();
+		this.pop_ignores();
 
 		this.fragment = new Fragment(this, ast.html);
 		this.name = this.get_unique_name(name);
 
+		this.push_ignores(this.ast.instance ? extract_ignores_above_position(this.ast.instance.start, this.ast.html.children) : []);
 		this.walk_instance_js_post_template();
+		this.pop_ignores();
 
 		this.elements.forEach(element => this.stylesheet.apply(element));
 		if (!compile_options.customElement) this.stylesheet.reify();
