+++ test/validator/samples/silence-warnings-2/input.svelte
@@ -1,0 +1,9 @@
+<script context="module">
+	let foo;
+</script>
+
+<!-- svelte-ignore unused-export-let module-script-reactive-declaration -->
+<script>
+	export let unused;
+	$: reactive = foo;
+</script>
