+++ test/validator/samples/silence-warnings-2/warnings.json
@@ -1,0 +1,1 @@
+[]
